% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_pie.R
\name{cloud_pie}
\alias{cloud_pie}
\title{Cloud Pie: A Totally Unnecessary Frankenstein's Monster.}
\usage{
cloud_pie(data, name_col, count_col, ...)
}
\arguments{
\item{data}{Summary dataframe with two columns: categories, and counts for those categories.}

\item{name_col}{Column containing the category name (character).}

\item{count_col}{Column containing the counts for each category (numeric).}
}
\description{
An unholy matrimony of a 3D exploded pie chart and a word cloud. You're welcome. Just promise you'll never, ever use this function.
}
\examples{
library(dplyr)

data <- read.csv("https://raw.githubusercontent.com/matt-dray/draytasets/master/ssb_pokeballs.csv")

pkmn <- data \%>\%
  group_by(pokemon) \%>\%
  count() \%>\%
  ungroup()

cloud_pie(data = pkmn, name_col = "pokemon", count_col = "n")
}
\keyword{ohgodmyeyes}
\keyword{pie,}
\keyword{wordcloud,}
