% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collembola_function.R
\name{collembola_function}
\alias{collembola_function}
\title{A Collembola Function}
\usage{
collembola_function(love = TRUE)
}
\arguments{
\item{love}{Do you love collembola? Defaults to TRUE.}
}
\description{
This function allows you to express your love of collembola.
}
\examples{
collembola_function()
}
\keyword{collembola}
