% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_kit.R
\name{create_brickr_player}
\alias{create_brickr_player}
\title{Create A Brick Soccer Player}
\usage{
create_brickr_player(hair_col = 39, skin_col = 8, boot_col = 2,
  shirt_body_col = 7, shirt_sleeve_col = 1, shorts_col = 1,
  sock_col = 7, sock_trim_col = 1)
}
\arguments{
\item{hair_col}{Numeric code as per \code{brickr::lego_colors}}

\item{skin_col}{Numeric code as per \code{brickr::lego_colors}}

\item{boot_col}{Numeric code as per \code{brickr::lego_colors}}

\item{shirt_body_col}{Numeric code as per \code{brickr::lego_colors}}

\item{shirt_sleeve_col}{Numeric code as per \code{brickr::lego_colors}}

\item{shorts_col}{Numeric code as per \code{brickr::lego_colors}}

\item{sock_col}{Numeric code as per \code{brickr::lego_colors}}

\item{sock_trim_col}{Numeric code as per \code{brickr::lego_colors}}
}
\description{
Create a tibble that defines a \{brickr\} soccer player. Pass to
\code{\link[brickr:bricks_from_table]{brickr::bricks_from_table()}} and then
\code{\link[brickr:display_bricks]{brickr::display_bricks()}} to render it.
}
\examples{
man_city <- create_brickr_player(
  hair_col = 2,
  skin_col = 39,
  boot_col = 7,
  shirt_body_col = 13,
  shirt_sleeve_col = 13,
  shorts_col = 1,
  sock_col = 27,
  sock_trim_col = 27
)

\dontrun{
library(dplyr)
library(brickr)
create_brickr_player() \%>\%
  bricks_from_table(lego_colors) \%>\%
  display_bricks()
}
}
\keyword{soccer}
