% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gif.R
\name{make_gif}
\alias{make_gif}
\title{Make a gif from a folder of PNG image files.}
\usage{
make_gif(png_folder = "img", gif_path = "new_gif.gif", ...)
}
\arguments{
\item{png_folder}{Folder where you've stored your PNG image files.}

\item{gif_path}{File path for your new gif, must end in '.gif' extension.}

\item{...}{Pass further arguments to gifski (see ?gifski for details).}
}
\description{
Point to a folder of PNG files and stitch them into a gif. Convenience function for the gifski function.
}
\examples{
# TODO
}
\keyword{gif,}
\keyword{gifski}
