% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phd_data.R
\docType{data}
\name{phd}
\alias{phd}
\title{PhD Thesis By Matt Dray.}
\format{An object of class \code{character} of length 1023.}
\source{
\href{http://orca.cf.ac.uk/id/eprint/68365/}{Orca online research portal}
}
\usage{
data(phd)
}
\description{
Text of the PhD thesis 'Effects of multiple environmental stressors on litter
chemical composition and decomposition' (Dray, 2014). Character vector where
each element is a line of text.
}
\examples{
data(phd)
}
\references{
Dray, Matthew William (2014). Effects of multiple environmental
stressors on litter chemical composition and decomposition. PhD thesis,
Cardiff University.
}
\keyword{datasets}
