% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkmn_go_data.R
\docType{data}
\name{pkmn_go}
\alias{pkmn_go}
\title{Pokemon Go Captures.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 696 rows and 9 columns.}
\source{
\href{https://www.pokemongo.com/}{Pokemon Go}
}
\usage{
data(pkmn_go)
}
\description{
Details of nearly 700 Pokemon captured in Autumn 2016 on the Pokemon Go
video game app on Apple iOS.
}
\examples{
data(pkmn_go)
}
\references{
Pokemon Go, developed and published by Niantic, with characters
developed by The Pokemon Company.
(\href{https://www.nianticlabs.com/}{Niantic})
}
\keyword{datasets}
