% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ye_cols.R
\docType{data}
\name{ye_cols}
\alias{ye_cols}
\title{A Vector Of Colours From The Ye Album Cover.}
\format{An object of class \code{character} of length 7.}
\usage{
ye_cols
}
\value{
A list of named hexadecimal colours.
}
\description{
Simplified colours sampled from the cover of Kanye West's 2018 album 'ye'.
}
\examples{
library(dray)
library(ggplot2)

   ggplot(mtcars) +
   aes(x = mpg, y = wt) +
   geom_point(
   colour = ye_cols["text_green"]
   )
}
\keyword{color,}
\keyword{colour,}
\keyword{kanye,}
\keyword{pallette,}
\keyword{ye}
