% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_pie.R
\name{cloud_pie}
\alias{cloud_pie}
\title{Cloud Pie: A Totally Unnecessary Frankenstein's Monster.}
\usage{
cloud_pie(data, name_col, count_col, ...)
}
\arguments{
\item{data}{Summary dataframe with two columns: categories, and counts for those categories.}

\item{name_col}{Column containing the category name.}

\item{name_col}{Column containing the counts for each category.}
}
\description{
An unholy matrimony of a 3D-Exploded-Pie-Chart and a word cloud. You're welcome.
}
\examples{
cat_function()
}
