% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gif.R
\name{make_gif}
\alias{make_gif}
\title{Make a gif from a folder of PNG image files.}
\usage{
make_gif(png_folder = "img", gif_path = "new_gif.gif", ...)
}
\arguments{
\item{png_folder}{Folder where you've stored your PNG image files (.png or .PNG).}

\item{gif_path}{File path for your new gif including .gif extension.}

\item{...}{Pass further arguments to gifski (see ?gifski for details).}
}
\description{
Point to a folder of PNG files and engifify them (make them into a gif (I made this word up)). A convenience function for the gifski function.
}
\examples{
\dontrun{
# filepath of folder containing PNGs
gif_in <- "folder/subfolder"

# filepath for the output gif (including extension)
gif_out <- "gifs/amazing_gif.gif"

# engifify
make_gif(
  png_folder = gif_in,
  gif_path = gif_out
)
}
}
\keyword{gif,}
\keyword{gifski}
